requirements.txt
langchain==0.1.16
langchain-community==0.0.33
langchain-openai==0.1.3
langchain-huggingface          # 新增：用於地端 Embedding
sentence-transformers          # 新增：用於地端 Embedding
faiss-cpu                      # 修改：移除 ==1.8.0，讓 Windows 自動找可用版本
llama-parse==0.4.2
llama-index-core==0.10.30
python-dotenv==1.0.1
nest_asyncio==1.6.0
neo4j